Account for un-revealed pokemon in the eval
This is to help monte-carlo search.

Previously, when pokemon were un-revealed the evaluation function would
treat the un-revealed pokemon as fainted. For example, on the first turn of a
randombattle when it sees 6v1 it would result in roughly a (5 x pkmn
value) advantage for side_one. This was not a problem for Expectiminimax
as the algorithm is trying to minimize the loss.

However, with Monte-Carlo search, a +500 value compared to a +400 value
can be seen as roughly equal after being passed through the sigmoid.
That is not acceptable since that difference of 100 points represents an
entire pokemon's worth of value.

This change adds bonus points to side_two if pokemon are un-revealed.
These values are the same as if the pokemon was revealed and at 100% hp,
which they basically are.