Remove HashMap from mcts Node
The HashMap was pretty easy to get things going but with only 13
options its faster to cast the MoveChoice to `usize` and use an
active flag to check whether or not it is a valid move

Results in about a 20-25% gain in number of nodes that can be searched, and
uses an equivalent amount of memory.

I also tried to remove the HashMap from the `Node.children` attribute, but
it ended up being slower and consuming much more memory. Might come back
to that later