Don't expand monte-carlo if both moves are None
If both moves are None then we have reached either the end of the game
or a situation where only unrevealed pokemon are remaining for side_two.
There is no reason in this case to expand because the resulting node
will be the same thing

Also, cut off search after 10M iterations. See comment for reasoning