perf: modify mcts rollout parameter to be the change in evaluation rather than the absolute evaluation
The sigmoid function is tuned so that a 200 point swing (about 2 full pokemon) is close to 0.0 or 1.0. In situations where the bot is very winning or very losing it can sometimes fail to see the difference between moves because they evaluate so similarly.

This results in a very slight gain in performance during self-play: 51% WR in about 3300 games.

with this change the evaluation function no longer needs an evaluation
boost for unrelealed pokemon