perf: MCTS Performance Improvements
Remove bloated SideOptions struct that inefficiently created a new
MoveNode 13 times, regardless of whether or not 13 were needed. This
also removes the need for the active attribute on the MoveNode. Rather
than indexing SideOptions by a MoveChoice, a Vec<MoveNode> is used and
it is indexed by a usize. The s1_choice and s2_choice attributes of
the Node are now usize rather than MoveChoice and
maximize_ucb_for_side will spit out a usize rather than a MoveChoice

Results in about a ~15% speed increase and approx 20-25% less memory
usage.