refactor: Remove last_used_move & damage_dealt features
Instead these features are always enabled but the engine checks at
run-time whether or not they should generate instructions. Since there
tends to be many of these instructions, not generating them results in about
a 20% increase in the number of monte-carlo iterations that can be done
which has a noticeable (52%) performance gain in 100ms self-play. This may not
be noticeable with 10s decisions.

Also decrement PP only if the move being used starts with less than 10 pp.
Another (micro?) optimization since if a move has more than 10pp then
it is (probably?) too early to start caring about PP. I may revert
this later but it will be hard to get data about whether or not this
makes a difference because pp decrementing doesn't show a noticeable
difference in battle strength for 100ms decisions