feat: Branch if damage is a range to kill
Adds a new feature to all generations (except gen1) that causes the
instruction generation to create a new branch if the damage being dealt
has a chance to kill and a chance to not kill. This is not added to gen1
because gen1 already branches in a similar way on critical hits due to
them being quite common.

This doesn't show a significant performance gain in self-play but there
are scenarios that I have observed where assuming the average damage
roll resulted in an endgame throw.

This feature is enabled during MCTS but only at the root node. I should
experiment with adding it to the first child as well. Beyond that, there
would be some serious performance hits - roughly 12% fewer MCTS
iterations if let run for the entire search.