feat: consider critical strikes
Consider critical strikes when there is no chance for an attack to kill.
If min damage is less than the target's HP a branch happens for crits.
This is mutually exclusive with the damage roll branch.

When considering a damage roll the crit chance is added to the branch
that kills

I thought this wouldn't have a change in performance but about
1,000 300ms self-play battles later and this gets about a 54% winrate.