fix: Check para/sleep/confusion before "before_move()"
The logic in before_move() assumes that you are going to successfully
attempt to execute your move. This means that if you are
fully-paralysed, staying asleep, staying frozen, hitting yourself in confusion we do NOT
want to run before_move(). This fixes a bug where a pokemon attempting
to use explosion would kill itself but then be fully-paralysed and do 0
damage. The branch for paralysis should happen before before_move() is
called.

There was one piece of logic in before_move() that did need to be
checked before the paralysis, sleep, and confusion checks - pinch
berries. Pinch berries in this engine are checked for both sides before
each attempted move, which is a hack to put it lightly. This
commit also moves the pinch berry checks to its own function and calls that
function before the paralysis/sleep/confusion/frozen check.