feat: Proper Side Condition Durations Tracking
Somehow I've gotten this far without proper tracking
of reflect/lightscreen/auroraveil/tailwind turns remaining. This
uses the i8 value of a side condition to represent the turns remaining
rather than "1" for being active and "0" for inactive. End of turn this
value is decremented if it is greater than 0. Existing logic already was
checking for > 0 for the side condition to be active so no changes
there.

Also account for aurora veil for the move raging bull