refactor: Damage calc returns crit damage as well
This is important because you can't just multiple actual damage by 1.5 since crits ignore some boosts as well as screens. Instead we'll let the engine tell us what the damages are for non-crit and crit damage

This also changes the damage calc logic to return 2 values for each side: the maximum non-crit damage and the maximum crit damage. The caller of this code can multiply by 0.85 to get the min roll if they please