fix: remove duplicated logic
a bunch of duplicated functions w.r.t. stringify-ing a movechoice. This brings them all together into 1 function with an override being made for the python bindings

Also moves a lot of the MoveChoice logic into the state file so that
things can be split up by generation a bit easier in future commits