refactor: move state structs and common functions to top level
Creates a src/state.rs to contain the common state code and leaves
anything generation-specific for that generation's state.rs. This allows
common logic to be common while also allowing per-generation overrides.

A lot of what makes a State is common across generations in this project.
Yes technically I can keep these separate per generation and even reduce
the size of the struct for earlier generations but I don't see the
point. I do not copy states in any hot-paths so earlier generations do
not need to optimize the state to be small. The reduction in duplicated
code is well worth it.