refactor: refactor end_of_turn_triggered
refactors the logic that checks for end of turn being triggered to only check for is either side was:
- replacing a fainted pkmn
- was forced to switch for some other reason (e.g. uturn pivot)

in both of these situations, end of turn is NOT triggered, otherwise it is. This simplified the logic and makes it more correct

This is more correct for gen9, and the logic is shared for gens 4-9 since this code is shared. I did not investigate differences between generations, and gen3 remains incorrect