refactor: Lazily Expand Nodes
Oh boy why didn't I do this initially.

Don't call get_all_options() for each child node as the nodes are
created, but instead wait until that node is selected to get all
options. This saves a LOT of memory, roughly 30%-50% in singles, since
we no longer waste resources populating vectors on leaf nodes that are
never selected. This also increases the efficiency of
monte-carlo search (maybe 20%) because we no longer waste time
generating options and creating/dropping vectors that are not selected.

To do this, a raw pointer is used for s1_options and s2_options and
checking for null is done on selection.